Runs on linux

+++ src/nn/optimizer/adam.h
@@ -1,7 +1,7 @@
 #pragma once
 
+#include "../../operations/operations.h"
 #include "optimizer.h"
-#include "operations/operations.h"
 
 namespace nn {
 
@@ -35,8 +35,8 @@ struct Adam : public Optimizer{
     }
 
     void step(OptimizerEntry& entry, int idx, float lr) override {
-        float bc1 = 1.0f - std::powf(beta1, step_);
-        float bc2 = 1.0f - std::powf(beta2, step_);
+        float bc1 = 1.0f - std::pow(beta1, step_);
+        float bc2 = 1.0f - std::pow(beta2, step_);
         float slr = lr * sqrtf(bc2) / bc1;
 
         operations::adam<data::GPU>(
